app/requirements.txt
fastapi
uvicorn
motor
pydantic
exrex
python-dotenv